M-71: refactor authenticated layout, add create room page and panel
