+++ app/Http/Controllers/LoginController.php
@@ -2,15 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
-use App\Models\User;
-
 class LoginController extends Controller
 {
-	public function authenticate(Request $request)
+	public function authenticate(Request $request): RedirectResponse
 	{
 		$credentials = $request->validate([
 			"email" => ["required", "email"],
