+++ app/Http/Controllers/LogoutController.php
@@ -2,16 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\Auth;
-
-use App\Http\Controllers\Controller;
-use App\Models\User;
 
 class LogoutController extends Controller
 {
-	public function logout(Request $request)
-	{
+	public function logout(
+		Request $request
+	): Redirector|Application|RedirectResponse {
 		Auth::logout();
 
 		$request->session()->invalidate();
