--- app/Http/Controllers/RoomsController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Inertia\Inertia;
-
-class RoomsController extends Controller
-{
-	public function index()
-	{
-		return Inertia::render("rooms");
-	}
-}
