+++ resources/js/components/FormField/FormField.tsx
@@ -19,14 +19,16 @@ export const FormField: React.FC<FormFieldInterface> = ({
 	children,
 }) => (
 	<div>
-		<label htmlFor={identifier}>{children}</label>
+		<label htmlFor={identifier} className="text-sm text-gray-900">
+			{children}
+		</label>
 		<input
 			id={identifier}
 			name={identifier}
 			type={type}
 			autoComplete={autocomplete}
 			required={required}
-			className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
+			className="appearance-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
 			value={value}
 			onChange={handleChange}
 		/>
